Shridbranch

+++ VMBackup/main/workloadPatch/WorkloadPatch.py
@@ -27,6 +27,7 @@ except ImportError:
     import configparser as ConfigParsers
 import subprocess
 from common import CommonVariables
+from workloadPatch.logbackupPatch import logbackup
 
 class ErrorDetail:
     def __init__(self, errorCode, errorMsg):
@@ -37,7 +38,7 @@ class WorkloadPatch:
     def __init__(self, logger):
         self.logger = logger
         self.name = "oracle"
-        self.command = "/usr/bin/"
+        self.command = "sqlplus"
         self.dbnames = []
         self.cred_string = ""
         self.ipc_folder = None
@@ -47,6 +48,7 @@ class WorkloadPatch:
         self.child = []
         self.timeout = 90
         self.outfile = ""
+        self.logbackup = ""
         self.confParser()
 
     def pre(self):
@@ -120,59 +122,59 @@ class WorkloadPatch:
             global preOracleStatus
             preOracleStatus = self.databaseStatus()
             if "OPEN" in str(preOracleStatus):
-                self.logger.log("Shrid: Pre- Database is open")
-                print("Shrid: Pre- Database is open")
+                self.logger.log("WorkloadPatch: Pre- Database is open")
+                print("WorkloadPatch: Pre- Database is open")
             else:
-                self.logger.log("Shrid: Pre- Database not open. Backup may proceed without pre and post")
-                print("Shrid: Pre- Database not open. Backup may proceed without pre and post")
+                self.logger.log("WorkloadPatch: Pre- Database not open. Backup may proceed without pre and post")
+                print("WorkloadPatch: Pre- Database not open. Backup may proceed without pre and post")
                 return None
 
-            print("Shrid: Pre- Inside oracle pre")
-            self.logger.log("Shrid: Pre- Inside oracle pre")
-            preOracle = "sqlplus -s / as sysdba @" + os.path.join(os.getcwd(), "main/workloadPatch/scripts/preOracleMaster.sql ")
+            print("WorkloadPatch: Pre- Inside oracle pre")
+            self.logger.log("WorkloadPatch: Pre- Inside oracle pre")
+            preOracle = self.command + " -s / as sysdba @" + os.path.join(os.getcwd(), "main/workloadPatch/scripts/preOracleMaster.sql ")
             args = ["su", "-", self.cred_string, "-c", preOracle]
             process = subprocess.Popen(args)
             while process.poll() == None:
                 sleep(1)
             self.timeoutDaemon()
-            self.logger.log("Shrid: Pre- Exiting pre mode for master")
-            print("Shrid: Pre- Exiting pre mode for master")
+            self.logger.log("WorkloadPatch: Pre- Exiting pre mode for master")
+            print("WorkloadPatch: Pre- Exiting pre mode for master")
         #----SHRID CODE END----#
 
     #----SHRID CODE START----#
     def timeoutDaemon(self):
         global preDaemonThread
         if 'oracle' in self.name.lower():
-            self.logger.log("Shrid: Inside oracle condition in timeout daemon")
-            print("Shrid: Inside oracle condition in timeout daemon")
-            preDaemonOracle = "sqlplus -s / as sysdba @" + os.path.join(os.getcwd(), "main/workloadPatch/scripts/preOracleDaemon.sql ") + self.timeout
+            self.logger.log("WorkloadPatch: Inside oracle condition in timeout daemon")
+            print("WorkloadPatch: Inside oracle condition in timeout daemon")
+            preDaemonOracle = self.command + " -s / as sysdba @" + os.path.join(os.getcwd(), "main/workloadPatch/scripts/preOracleDaemon.sql ") + self.timeout
             argsDaemon = ["su", "-", self.cred_string, "-c", preDaemonOracle]
             preDaemonThread = threading.Thread(target=self.threadForTimeoutDaemon, args=[argsDaemon])
             preDaemonThread.start()
-        self.logger.log("Shrid: timeoutDaemon started for: " + self.timeout + " seconds")
-        print("Shrid: timeoutDaemon started for: ", self.timeout, " seconds")
+        self.logger.log("WorkloadPatch: timeoutDaemon started for: " + self.timeout + " seconds")
+        print("WorkloadPatch: timeoutDaemon started for: ", self.timeout, " seconds")
     #----SHRID CODE END----#
 
     #----SHRID CODE START----#
     def threadForTimeoutDaemon(self, args): 
             global daemonProcess
             daemonProcess = subprocess.Popen(args)
-            self.logger.log("Shrid: daemonProcess started")
-            print("Shrid: daemonProcess started")
+            self.logger.log("WorkloadPatch: daemonProcess started")
+            print("WorkloadPatch: daemonProcess started")
             while daemonProcess.poll() == None:
                 sleep(1)
-            self.logger.log("Shrid: daemonProcess completed")
-            print("Shrid: daemonProcess completed")
+            self.logger.log("WorkloadPatch: daemonProcess completed")
+            print("WorkloadPatch: daemonProcess completed")
     #----SHRID CODE END----#
 
     #---- SHRID CODE START----#
     def databaseStatus(self):
 
         if 'oracle' in self.name.lower():
-            statusArgs =  "su - " + self.cred_string + " -c " + "'sqlplus -s / as sysdba<<-EOF\nSELECT STATUS FROM V\$INSTANCE;\nEOF'"
+            statusArgs =  "su - " + self.cred_string + " -c " + "'" + self.command +" -s / as sysdba<<-EOF\nSELECT STATUS FROM V\$INSTANCE;\nEOF'"
             oracleStatus = subprocess.check_output(statusArgs, shell=True)
-            self.logger.log("Shrid: databaseStatus- " + str(oracleStatus))
-            print("Shrid: databaseStatus- ", str(oracleStatus))
+            self.logger.log("WorkloadPatch: databaseStatus- " + str(oracleStatus))
+            print("WorkloadPatch: databaseStatus- ", str(oracleStatus))
             return oracleStatus
 
         return False
@@ -198,16 +200,18 @@ class WorkloadPatch:
         
     def preSlave(self):
         self.logger.log("WorkloadPatch: Entering post mode for master")
-        if os.path.exists(self.outfile):
-            os.remove(self.outfile)
-        else:
-            self.logger.log("WorkloadPatch: File for IPC does not exist at post")
-        if len(self.child) == 0:
-            self.logger.log("WorkloadPatch: Not app consistent backup")
-            self.error_details.append("not app consistent")
-        elif self.child[0].poll() is None:
-            self.logger.log("WorkloadPatch: pre connection still running. Sending kill signal")
-            self.child[0].kill()
+        if self.ipc_folder != None:
+            if os.path.exists(self.outfile):
+                os.remove(self.outfile)
+            else:
+                self.logger.log("WorkloadPatch: File for IPC does not exist at post")
+            if len(self.child) == 0:
+                self.logger.log("WorkloadPatch: Not app consistent backup")
+                self.error_details.append("not app consistent")
+            elif self.child[0].poll() is None:
+                self.logger.log("WorkloadPatch: pre connection still running. Sending kill signal")
+                self.child[0].kill()
+        
         if 'mysql' in self.name.lower():
             self.logger.log("WorkloadPatch: Create connection string for post master")
             args = "sudo "+self.command+self.name+" --login-path="+self.cred_string+" < main/workloadPatch/scripts/postMysqlSlave.sql"
@@ -218,16 +222,18 @@ class WorkloadPatch:
     
     def postMaster(self):
         self.logger.log("WorkloadPatch: Entering post mode for master")
-        if os.path.exists(self.outfile):
-            os.remove(self.outfile)
-        else:
-            self.logger.log("WorkloadPatch: File for IPC does not exist at post")
-        if len(self.child) == 0:
-            self.logger.log("WorkloadPatch: Not app consistent backup")
-            self.error_details.append("not app consistent")
-        elif self.child[0].poll() is None:
-            self.logger.log("WorkloadPatch: pre connection still running. Sending kill signal")
-            self.child[0].kill()
+        if self.ipc_folder != None:
+            if os.path.exists(self.outfile):
+                os.remove(self.outfile)
+            else:
+                self.logger.log("WorkloadPatch: File for IPC does not exist at post")
+            if len(self.child) == 0:
+                self.logger.log("WorkloadPatch: Not app consistent backup")
+                self.error_details.append("not app consistent")
+            elif self.child[0].poll() is None:
+                self.logger.log("WorkloadPatch: pre connection still running. Sending kill signal")
+                self.child[0].kill()
+        
         if 'mysql' in self.name.lower():
             if len(self.child) == 0:
                 self.logger.log("WorkloadPatch: Not app consistent backup")
@@ -241,14 +247,14 @@ class WorkloadPatch:
         elif 'oracle' in self.name.lower():
             postOracleStatus = self.databaseStatus()
             if postOracleStatus != preOracleStatus:
-                self.logger.log("Shrid: Error. Pre and post database status different.")
-                print("Shrid: Error. Pre and post database status different.")
+                self.logger.log("WorkloadPatch: Error. Pre and post database status different.")
+                print("WorkloadPatch: Error. Pre and post database status different.")
             if "OPEN" in str(postOracleStatus):
-                self.logger.log("Shrid: Post- Database is open")
-                print("Shrid: Post- Database is open")
+                self.logger.log("WorkloadPatch: Post- Database is open")
+                print("WorkloadPatch: Post- Database is open")
             else:
-                self.logger.log("Shrid: Post- Database not open. Backup may proceed without pre and post")
-                print("Shrid: Post- Database not open. Backup may proceed without pre and post")
+                self.logger.log("WorkloadPatch: Post- Database not open. Backup may proceed without pre and post")
+                print("WorkloadPatch: Post- Database not open. Backup may proceed without pre and post")
                 return
 
             self.logger.log("Shird: Post- Inside oracle post")
@@ -256,20 +262,21 @@ class WorkloadPatch:
             if preDaemonThread.isAlive():
                 self.logger.log("Shird: Post- Timeout daemon still in sleep")
                 print("Shird: Post- Timeout daemon still in sleep")
-                self.logger.log("Shrid: Post- Initiating Post Script")
-                print("Shrid: Post- Initiating Post Script")
+                self.logger.log("WorkloadPatch: Post- Initiating Post Script")
+                print("WorkloadPatch: Post- Initiating Post Script")
                 daemonProcess.terminate()
             else:
-                self.logger.log("Shrid: Post error- Timeout daemon executed before post")
-                print("Shrid: Post error- Timeout daemon executed before post")
+                self.logger.log("WorkloadPatch: Post error- Timeout daemon executed before post")
+                print("WorkloadPatch: Post error- Timeout daemon executed before post")
                 return
-            postOracle="sqlplus -s / as sysdba @" + os.path.join(os.getcwd(), "main/workloadPatch/scripts/postOracleMaster.sql ")
+            postOracle = self.command + " -s / as sysdba @" + os.path.join(os.getcwd(), "main/workloadPatch/scripts/postOracleMaster.sql ")
             args = ["su", "-", self.cred_string, "-c", postOracle]
             process = subprocess.Popen(args)
             while process.poll()==None:
                 sleep(1)
-            self.logger.log("Shrid: Post- Completed")
-            print("Shrid: Post- Completed")
+            self.logger.log("WorkloadPatch: Post- Completed")
+            print("WorkloadPatch: Post- Completed")
+            self.callLogbackup()
         #----SHRID CODE END----#
 
     def postMasterDB(self):
@@ -316,6 +323,9 @@ class WorkloadPatch:
                     if config.has_option("workload", 'dbnames'):
                         dbnames_list = config.get("workload", 'dbnames') #mydb1;mydb2;mydb3
                         self.dbnames = dbnames_list.split(';')
+                    if config.has_section("logbackup"):
+                        self.logbackup = "enable"
+                        self.logger.log("WorkloadPatch: Logbackup Enabled")
                 else:
                     self.error_details.append(ErrorDetail(CommonVariables.FailedPreWorkloadPatch, "no matching workload config found"))
             else:
@@ -334,66 +344,11 @@ class WorkloadPatch:
     
     def getRole(self):
         return "master"
-
-#----SHRID CODE START----#
-class logbackup:
-    def __init__(self):
-        self.name = "oracle"
-        self.cred_string = "AzureBackup"
-        self.baseLocation = "/hdd/AutoIncrement/"
-        self.parameterFilePath = "/u01/app/oracle/product/19.3.0/dbhome_1/dbs/initCDB1.ora"
-        self.oracleParameter = {}
-        self.backupSource = ""
-        self.crontabLocation = "/var/spool/cron/root"
-        self.confParser()
-        self.crontabEntry()
-
-    def crontabEntry(self):
-        if os.path.exists(self.crontabLocation):
-            crontabFile = open(self.crontabLocation, 'r')
-            crontabCheck = crontabFile.read()
-        else:
-            crontabCheck = "NO CRONTAB"
-
-        if 'oracle' in self.name.lower():
-            if 'logbackup' in str(crontabCheck):
-                print("logbackup: Existing Crontab Entry: ", str(crontabCheck))
-                return
-            else:
-                os.system("echo \"*/15 * * * * python " + os.path.join(os.getcwd(), "main/logbackup.py\"") + " >> /var/spool/cron/root")
-                print("logbackup: New Crontab Entry Made")
-                return
     
-    def confParser(self):
-        print("#---- WorkloadPatch: Entering workload config parsing ----#")
-        configfile = '/etc/azure/workload.conf' 
-        if os.path.exists(configfile):
-            config = ConfigParsers.ConfigParser()
-            config.read(configfile)
-            if config.has_section("logbackup"):
-                print("    logbackup: config section present for logbackup ")
-                if config.has_option("logbackup", 'workload_name'):                        
-                    self.name = config.get("logbackup", 'workload_name')
-                    print("    logbackup: config logbackup workload name: ", self.name)
-                else:
-                    return None
-                if config.has_option("logbackup", 'loginPath'):
-                    self.cred_string = config.get("logbackup", 'loginPath')
-                    print("    logbackup: config logbackup credential string: ", self.cred_string)
-                if config.has_option("logbackup", 'parameterFilePath'):
-                    self.parameterFilePath = config.get("logbackup", 'parameterFilePath')
-                    print("    logbackup: config logbackup parameter file path: ", self.parameterFilePath)
-                if config.has_option("logbackup", 'baseLocation'):
-                    self.baseLocation = config.get("logbackup", 'baseLocation')
-                    print("    logbackup: config logbackup base location: ", self.baseLocation)
-                if config.has_option("logbackup", 'backupSource'):
-                    self.backupSource = config.get("logbackup", 'backupSource')
-                    print("    logbackup: config logbackup backup source: ", self.backupSource)
-                if config.has_option("logbackup", 'crontabLocation'):
-                    self.crontabLocation = config.get("logbackup", 'crontabLocation')
-                    print("    logbackup: config logbackup crontab location: ", self.crontabLocation)
-                print("#----End of Config Parser ----#")
+    def callLogbackup(self):
+        if 'enable' in self.logbackup.lower():
+            self.logger.log("WorkloadPatch: Initializing logbackup")
+            print("WorkloadPatch: Initializing logbackup")
+            logbackupObject = logbackup()
         else:
-            print("No matching workload config found")
-
-#----SHRID CODE END----#
+            return
